LayoutTests/fast/dom/Window/setInterval-setTimeout-zero-ordering-expected.txt

Checks that setTimeout(0) & setInterval(0) ordering is maintained

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


* setInterval(0)
* setTimeout(0)
PASS setInterval0Called is true
* setInterval(1)
PASS setInterval0Called is true
PASS setTimeout0Called is true
* setTimeout(1)
PASS setInterval1Called is true
PASS successfullyParsed is true

TEST COMPLETE
